+++ apps/desktop/src/lib/chat-messages/index.ts
@@ -1,0 +1,18 @@
+export { toChatMessages } from './hydration'
+export {
+  appendAssistantTextPart,
+  appendReasoningPart,
+  assistantTextPart,
+  chatMessageText,
+  collectUnspokenTurnSpeech,
+  completeOpenTimelineParts,
+  dedupeRepeatedTextInParts,
+  mergeFinalAssistantText,
+  reasoningPart,
+  renderMediaTags,
+  textPart
+} from './parts'
+export type { UnspokenTurnSpeech } from './parts'
+export { branchGroupForUser, preserveLocalAssistantErrors } from './reconciliation'
+export { sealOpenToolParts, upsertToolPart } from './tool-parts'
+export type { ChatMessage, ChatMessagePart, GatewayEventPayload, TimelinePartMetadata } from './types'
